// src/api/voicechat/model/chat-start-request-llm-config.ts
/* tslint:disable */
/* eslint-disable */
/**
 * voicechat
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface ChatStartRequestLlmConfig {
    'apiBaseUrl': string;
    'apiKey': string;
    'model': string;
}