// src/api/llm/model/config-list-my-config-request-query-filter.ts
/* tslint:disable */
/* eslint-disable */
/**
 * llm
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface ConfigListMyConfigRequestQueryFilter {
    'apiKey': string;
    'baseUrl': string;
    'contextLength': number;
    'description': string;
    'enableSearch': number;
    'enableThinking': number;
    'id': number;
    'maxTokens': number;
    'model': string;
    'name': string;
    'presencePenalty': number;
    'repetitionPenalty': number;
    'seed': number;
    'stream': number;
    'temperature': number;
    'topK': number;
    'topP': number;
    'userId': number;
}